fix(bench): isolate gate predicate from kysely-cursor import
Root typecheck runs before build and followed the new unit test into
bench/src/types.ts, which imports the package name. Move the pure gate
into bench/src/gate.ts (config-only deps) so tsc no longer needs dist/.
