+++ bench/src/compare.ts
@@ -1,32 +1,9 @@
-import type { DialectName } from './config.js'
 import type { BaselineCell, BaselineReport, CellDelta, CompareResult } from './types.js'
 import { cellKey } from './baseline.js'
+import { DEFAULT_REGRESSION_THRESHOLD, isGatingRegression, MIN_ABS_MS, MIN_GATE_DEPTH } from './gate.js'
 import { formatMs, formatSpeedup } from './metrics.js'
 
-export const DEFAULT_REGRESSION_THRESHOLD = 1.5
-
-/**
- * Shallow depths and sub-ms cells are dominated by runner noise on GHA.
- * Only enforce fail-on-regression for library pages at this depth and deeper.
- */
-export const MIN_GATE_DEPTH = 100
-
-/**
- * Minimum absolute cursor-mean slowdown (ms) required to fail CI, in addition
- * to the ratio threshold. Ratio-only spikes on flat sub-ms cells (e.g. +0.28ms
- * at 1.73×) are runner noise, not library regressions.
- *
- * Remote dialects (Docker RTT on GHA) use 2ms; in-process sqlite uses 0.5ms.
- */
-export const MIN_ABS_MS: Record<DialectName, number> = {
-  postgres: 2,
-  mysql: 2,
-  mssql: 2,
-  sqlite: 0.5,
-}
-
-/** Raw SQL ceiling — informative, not a library regression signal. */
-const NON_GATING_SCENARIOS = new Set(['ideal-baseline'])
+export { cursorDeltaMs, DEFAULT_REGRESSION_THRESHOLD, isGatingRegression, MIN_ABS_MS, MIN_GATE_DEPTH } from './gate.js'
 
 const safeRatio = (current: number, baseline: number): number => {
   if (!Number.isFinite(current) || !Number.isFinite(baseline) || baseline <= 0) {
@@ -40,35 +17,6 @@ const statusFor = (cursorRatio: number, threshold: number): CellDelta['status'] 
   if (cursorRatio >= threshold) return 'regression'
   if (cursorRatio <= 1 / threshold) return 'improvement'
   return 'stable'
-}
-
-/** Absolute cursor-mean delta in ms (current − baseline). */
-export const cursorDeltaMs = (d: CellDelta): number => d.current.cursorMean - d.baseline.cursorMean
-
-/**
- * Whether a cell regression should fail CI (`--fail-on-regression`).
- * Still listed in the report when true or false; only gating ones set exit 1.
- *
- * Fail only when all hold:
- * 1. status is regression (ratio ≥ threshold)
- * 2. library scenario (not ideal-baseline)
- * 3. depth ≥ MIN_GATE_DEPTH or walk label
- * 4. absolute delta ≥ MIN_ABS_MS[dialect]
- */
-export const isGatingRegression = (d: CellDelta): boolean => {
-  if (d.status !== 'regression') return false
-  if (NON_GATING_SCENARIOS.has(d.scenario)) return false
-  if (cursorDeltaMs(d) < MIN_ABS_MS[d.dialect]) return false
-
-  const depthMatch = /^depth=(\d+)$/.exec(d.label)
-  if (depthMatch) {
-    return Number(depthMatch[1]) >= MIN_GATE_DEPTH
-  }
-
-  // sequential-walk / author walk labels (`walk=N`) — gate these.
-  if (d.label.startsWith('walk=')) return true
-
-  return true
 }
 
 export const gatingRegressions = (result: CompareResult): CellDelta[] => result.regressions.filter(isGatingRegression)
